fix(tui): register sidecar toggle globally so tab works from home
The binding previously lived in the session route component, so tab
did nothing until the first message created a session. Extract the
toggle into src/sidecar.ts and register the command at app level,
falling back to a hint toast when there is no conversation yet.
